+++ pkg/ratelimit/resource.go
@@ -20,7 +20,10 @@ import (
 	"net"
 	"net/http"
 
+	"github.com/gorilla/mux"
+	"go.thethings.network/lorawan-stack/v3/pkg/auth"
 	"go.thethings.network/lorawan-stack/v3/pkg/events"
+	"go.thethings.network/lorawan-stack/v3/pkg/rpcmetadata"
 	"go.thethings.network/lorawan-stack/v3/pkg/ttnpb"
 	"go.thethings.network/lorawan-stack/v3/pkg/unique"
 )
@@ -44,11 +47,40 @@ type resource struct {
 func (r *resource) Key() string       { return r.key }
 func (r *resource) Classes() []string { return r.classes }
 
+const unauthenticated = "unauthenticated"
+
+func authTokenID(ctx context.Context) string {
+	if authValue := rpcmetadata.FromIncomingContext(ctx).AuthValue; authValue != "" {
+		_, id, _, err := auth.SplitToken(authValue)
+		if err != nil {
+			return unauthenticated
+		}
+		return id
+	}
+	return unauthenticated
+}
+
+var pathTemplate = func(r *http.Request) (string, bool) {
+	route := mux.CurrentRoute(r)
+	if route == nil {
+		return "", false
+	}
+	pathTemplate, err := route.GetPathTemplate()
+	if err != nil {
+		return "", false
+	}
+	return pathTemplate, true
+}
+
 // httpRequestResource represents an HTTP request. Avoid using directly, use HTTPMiddleware instead.
 func httpRequestResource(r *http.Request, class string) Resource {
+	specificClasses := make([]string, 0, 3)
+	if template, ok := pathTemplate(r); ok {
+		specificClasses = append(specificClasses, fmt.Sprintf("%s:%s", class, template))
+	}
 	return &resource{
 		key:     fmt.Sprintf("%s:ip:%s:url:%s", class, httpRemoteIP(r), r.URL.Path),
-		classes: []string{class, "http"},
+		classes: append(specificClasses, class, "http"),
 	}
 }