+++ CHANGELOG.md
@@ -12,6 +12,7 @@ For details about compatibility between different releases, see the **Commitment
 ### Added
 
 ### Changed
+
 - Server side events replaced with single socket connection using the native WebSocket API.
 
 ### Deprecated