+++ CHANGELOG.md
@@ -11,6 +11,8 @@ For details about compatibility between different releases, see the **Commitment
 
 ### Added
 
+- Rate limiting classes for individual HTTP paths.
+
 ### Changed
 
 ### Deprecated