+++ CHANGELOG.md
@@ -21,7 +21,7 @@ For details about compatibility between different releases, see the **Commitment
 
 ### Security
 
-## [3.28.1] - unreleased
+## [3.28.1] - 2023-11-27
 
 ### Added